adding payment page viewModel et Controller for strip

+++ GameGenesisApp/GameGenesisApp/Models/Product.cs
@@ -10,7 +10,7 @@ namespace GameGenesisApp.Models
         public string name { get; set; } = "Name";
         public string Description { get; set; } = "Description";
         public User Author { get; set; }
-        public float Price { get; set; }
+        public long Price { get; set; }
         public List<Category> Categories { get; set; }
         public List<Image> Images { get; set; }
     }
